+++ CHANGELOG.md
@@ -7,6 +7,9 @@ https://github.com/coleifer/peewee/releases
 
 ## master
 
+* Add `MySQLJSONField` (`playhouse.mysql_ext`) with `contains_any()` for the
+  `JSON_OVERLAPS`/"match any" counterpart to `contains` for JSON arrays.
+
 [View commits](https://github.com/coleifer/peewee/compare/4.1.0...master)
 
 ## 4.1.0